// src/6_shared/api/api-response.types.ts
export type ProductReview = {
    rating: number;
    comment: string;
    date: string;
    reviewerName: string;
    reviewerEmail: string;
};

export type ProductMeta = {
    createdAt: string;
    updatedAt: string;
    barcode: string;
    qrCode: string;
};

export type ProductDimensions = {
    width: number;
    height: number;
    depth: number;
};

export type ProductResponseFull = {
    id: number;
    title: string;
    description: string;
    category: string;
    price: number;
    discountPercentage: number;
    rating: number;
    stock: number;
    tags: string[];
    brand: string;
    sku: string;
    weight: number;
    dimensions: ProductDimensions;
    warrantyInformation: string;
    shippingInformation: string;
    availabilityStatus: string;
    reviews: ProductReview[];
    returnPolicy: string;
    minimumOrderQuantity: number;
    meta: ProductMeta;
    images: string[];
    thumbnail: string;
};

export type ProductsResponseFull ={
    products: ProductResponseFull[]
} 